fix: update tree depth uint type
Former-commit-id: 0f0fc95832d3b3b663d92f1d22f19a0e3a86ed42

+++ contracts/base/SemaphoreGroups.sol
@@ -21,7 +21,7 @@ abstract contract SemaphoreGroups is Context, ISemaphoreGroups {
     /// @param zeroValue: Zero value of the tree.
     function _createGroup(
         uint256 groupId,
-        uint8 depth,
+        uint256 depth,
         uint256 zeroValue
     ) internal virtual {
         if (groupId >= SNARK_SCALAR_FIELD) {
@@ -81,7 +81,7 @@ abstract contract SemaphoreGroups is Context, ISemaphoreGroups {
     }
 
     /// @dev See {ISemaphoreGroups-getDepth}.
-    function getDepth(uint256 groupId) public view virtual override returns (uint8) {
+    function getDepth(uint256 groupId) public view virtual override returns (uint256) {
         return groups[groupId].depth;
     }
